+++ R/class_build.R
@@ -11,10 +11,15 @@ build_init <- function(
     NULL
   }
   capture_warning <- function(condition) {
-    state$warnings <- paste(
-      c(state$warnings, build_message(condition)),
-      collapse = ". "
-    )
+    state$count_warnings <- (state$count_warnings %||% 0L) + 1L
+    should_store_warning <- (state$count_warnings < 51L) &&
+      (nchar(state$warnings %||% "") < build_message_max_nchar)
+    if (should_store_warning) {
+      state$warnings <- paste(
+        c(state$warnings, build_message(condition)),
+        collapse = ". "
+      )
+    }
     warning(as_immediate_condition(condition))
     invokeRestart("muffleWarning")
   }
@@ -89,7 +94,7 @@ build_message <- function(condition, prefix = character(0)) {
   out <- substr(
     paste(c(prefix, conditionMessage(condition)), collapse = " "),
     start = 0L,
-    stop = 2048L
+    stop = build_message_max_nchar
   )
   if_any(nzchar(out), out, ".")
 }
@@ -98,3 +103,5 @@ build_validate <- function(build) {
   tar_assert_correct_fields(build, build_new)
   metrics_validate(build$metrics)
 }
+
+build_message_max_nchar <- 2048L