+++ NEWS.md
@@ -1,4 +1,4 @@
-# targets 0.14.0.9000
+# targets 0.14.1
 
 * Fix the command inserted for debug mode (#975).
 * Set empty chunk options to ensure Target Markdown compatibility with the special "setup" chunk (#973, @KaiAragaki).